+++ tools/mkdocs_macros_bids/__init__.py
@@ -1,0 +1,19 @@
+from .main import define_env
+from .macros import (
+    make_filename_template,
+    make_entity_table,
+    make_entity_definitions,
+    make_suffix_table,
+    make_metadata_table,
+    make_filetree_example,
+)
+
+__all__ = [
+    "define_env",
+    "make_filename_template",
+    "make_entity_table",
+    "make_entity_definitions",
+    "make_suffix_table",
+    "make_metadata_table",
+    "make_filetree_example",
+]
