--- tools/mkdocs_macros_bidsschema/__init__.py
@@ -1,3 +1,0 @@
-from .main import define_env
-
-__all__ = ["define_env"]
