+++ tools/schemacode/__init__.py
@@ -1,8 +1,7 @@
 """A Python package for working with the BIDS schema."""
-from . import macros, schema, utils
+from . import schema, utils
 
 __all__ = [
-    "macros",
     "schema",
     "utils",
 ]
